% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_data.R
\docType{data}
\name{fish_species}
\alias{fish_species}
\title{Fish Species}
\format{
A \code{\link[tibble:tibble]{tibble::tibble()}} with 34 rows and 3 variables:
\describe{
\item{class}{class of fish (e.g. Panfish, Bass, Trout)}
\item{abbreviation}{abbreviation used in JSON (e.g. BLC = Black Crappie, WAE = Walleye)}
\item{name}{common name of fish (e.g. Bluegill, Rock Bass, Sauger)}
}
}
\source{
\url{https://maps2.dnr.state.mn.us/lakefinder/lib/lakefinder.js}
}
\usage{
fish_species
}
\description{
A \code{\link[tibble:tibble]{tibble::tibble()}} containing all fish species.
}
\keyword{datasets}
