% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_fishrmn.R
\docType{package}
\name{fishrmn-package}
\alias{fishrmn}
\alias{fishrmn-package}
\title{fishrmn:  Package for Minnesota fishermen}
\description{
R Package that interfaces with the MN DNR JSON API
}
\author{
\strong{Maintainer}: Aaron Anderson \email{aaron.m.anderson.3@gmail.com}

}
