% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_gazetteer.R
\name{gazetteer}
\alias{gazetteer}
\title{MN DNR Gazetteer}
\usage{
gazetteer(...)
}
\arguments{
\item{...}{Arguments to the JSON query.  \describe{
\item{name}{
Name of the object to retrieve
* line 1
* line 2
}
\item{type}{
Type of the object to retrieve.
* line 1
* line 2
}
}

county, lake, stprk, map100k, map24k, sna, nwr, wma, stream}
}
\description{
Wrapper to the \href{http://services.dnr.state.mn.us/api/gazetteer/v1/usage.html}{MN DNR Gazetteer API}.
}
