% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_utils.R
\name{json_handle_dots}
\alias{json_handle_dots}
\title{Handle ellipses for JSON queries}
\usage{
json_handle_dots(..., allowed_args = NULL, empty_allowed = FALSE)
}
\arguments{
\item{...}{Arguments passed on from query wrapper functions}

\item{allowed_args}{Names of the arguments that are allowed (\code{NULL} to allow
all arguments)}

\item{empty_allowed}{Can \code{...} be empty?}
}
\value{
\code{json_handle_dots()} returns a string that concatenates the names and values of the supplied
arguments.
}
\description{
Automatically transforms the ellipses arguments into a JSON query for the MN
DNR APIs.
}
\details{
Function will discard named arguments that are not allowed (determined by
\code{allowed_args}).  If a single unnamed argument is supplied with a single
\code{allowed_arg}, it will be automatically renamed.
}
\examples{
# normal execution
fishrmn:::json_handle_dots(name = "Waconia", type = "lake")
fishrmn:::json_handle_dots(name = c("Waconia", "Hydes"), type = "lake")

# discarded args
fishrmn:::json_handle_dots(name = c("Waconia", "Hydes"), type = "lake", allowed_args = "name")

# error, empty but not allowed
try(fishrmn:::json_handle_dots(), TRUE)
}
\keyword{internal}
