% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_utils.R
\name{try_read_json}
\alias{try_read_json}
\title{Error-Tolerant JSON Reader}
\usage{
try_read_json(url)
}
\arguments{
\item{url}{URL to read}
}
\value{
Parsed JSON list or data.frame
}
\description{
\code{\link[jsonlite:read_json]{jsonlite::read_json()}} with a \code{\link[rlang:try_fetch]{rlang::try_fetch()}} wrapper
}
\details{
If \code{\link[jsonlite:read_json]{jsonlite::read_json()}} is not successful, \code{\link[rlang:try_fetch]{rlang::try_fetch()}} contains
the error and a more informative error is returned.
}
\examples{
fishrmn:::try_read_json("http://services.dnr.state.mn.us/api/lakefinder/by_id/v1/?id=10005900")
fishrmn:::try_read_json("http://services.dnr.state.mn.us/api/lakefinder/by_id/v1/?")
}
\seealso{
\code{\link[jsonlite:read_json]{jsonlite::read_json()}} for JSON reading/parsing,
\code{\link[rlang:try_fetch]{rlang::try_fetch()}} for error handling.
}
\keyword{internal}
